pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools>=61.2",
    "setuptools_scm",
]

[project]
authors = [{name = "Common Partial Wave Analysis", email = "compwa-admin@ep1.rub.de"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "graphviz",
    "ipympl",
    "matplotlib <3.9.0", # https://github.com/ComPWA/compwa.github.io/actions/runs/9189453016/job/25271564131?pr=274
    "matplotlib",
    "mpl-interactions",
]
description = "Source code for the ComPWA Organization pages"
dynamic = ["version"]
license = {file = "LICENSE"}
maintainers = [{email = "compwa-admin@ep1.rub.de"}]
name = "compwa-org"
requires-python = ">=3.7"

[project.optional-dependencies]
dev = [
    "compwa-org[doc]",
    "compwa-org[jupyter]",
    "compwa-org[sty]",
    "compwa-org[test]",
    "sphinx-autobuild",
    "tox >=1.9", # for skip_install, use_develop
]
doc = [
    "Sphinx >=3",
    "myst-nb >=0.14", # nb_ configuration prefixes
    "sphinx-api-relink >=0.0.4",
    "sphinx-book-theme",
    "sphinx-codeautolink[ipython]",
    "sphinx-comments",
    "sphinx-copybutton",
    "sphinx-design",
    "sphinx-hep-pdgref",
    "sphinx-pybtex-etal-style",
    "sphinx-remove-toctrees",
    "sphinx-thebe",
    "sphinx-togglebutton",
    "sphinxcontrib-bibtex >=2",
]
jupyter = [
    "black",
    "isort",
    "jupyterlab",
    "jupyterlab-code-formatter",
    "jupyterlab-git",
    "jupyterlab-lsp",
    "jupyterlab-myst",
    "python-lsp-ruff",
    "python-lsp-server[rope]",
    'ypy-websocket <0.8.3; python_version <"3.8.0"',
]
sty = [
    "attrs",
    "compwa-org[types]",
    "nbformat",
    "pre-commit >=1.4.0",
    "ruff",
]
test = [
    "jupyter",
    "nbmake",
    'nbmake <1.3; python_version=="3.7.*"',
]
types = [
    "sphinx-api-relink >=0.0.4",
]

[project.readme]
content-type = "text/markdown"
file = "README.md"

[project.urls]
Documentation = "https://compwa.github.io"
Source = "https://github.com/ComPWA/compwa.github.io"
Tracker = "https://github.com/ComPWA/compwa.github.io/issues"

[tool.setuptools]
include-package-data = false

[tool.setuptools_scm]
write_to = "version.py"

[tool.pyright]
exclude = [
    "**/.git",
    "**/.ipynb_checkpoints",
    "**/.mypy_cache",
    "**/.pytest_cache",
    "**/.tox",
    "**/__pycache__",
    "**/_build",
]
reportGeneralTypeIssues = false
reportIncompatibleMethodOverride = false
reportMissingParameterType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportOperatorIssue = false
reportOverlappingOverload = false
reportPrivateImportUsage = false
reportPrivateUsage = false
reportReturnType = false
reportUnboundVariable = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryComparison = false
reportUnnecessaryContains = false
reportUnnecessaryIsInstance = false
reportUntypedFunctionDecorator = false
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedImport = true
reportUnusedVariable = true
typeCheckingMode = "strict"

[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--doctest-continue-on-failure",
    "--doctest-modules",
    "--durations=0",
    "--ignore-glob=*/.ipynb_checkpoints/*",
    "--ignore=docs/adr/001/sympy.ipynb",
    "--ignore=docs/conf.py",
    "--nbmake-timeout=1200",
    "-k not docs/report/019.ipynb",
]
filterwarnings = [
    "error",
    "ignore:Passing a schema to Validator.iter_errors is deprecated.*:DeprecationWarning",
    "ignore:The .* argument to NotebookFile is deprecated.*:pytest.PytestRemovedIn8Warning",
    "ignore:open_text is deprecated.*:DeprecationWarning",
    "ignore:read_text is deprecated.*:DeprecationWarning",
    "ignore:unclosed .*:ResourceWarning",
]
norecursedirs = [
    ".ipynb_checkpoints",
    ".virtual_documents",
    "_build",
    "docs/adr",
]
testpaths = [
    ".pre-commit",
    "docs",
]

[tool.ruff]
extend-include = ["*.ipynb"]
preview = true
show-fixes = true
target-version = "py37"

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[tool.ruff.lint]
extend-select = [
    "A",
    "B",
    "BLE",
    "C4",
    "C90",
    "D",
    "EM",
    "ERA",
    "FA",
    "I",
    "ICN",
    "INP",
    "ISC",
    "N",
    "NPY",
    "PGH",
    "PIE",
    "PL",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
ignore = [
    "D101",
    "D102",
    "D103",
    "D105",
    "D107",
    "D203",
    "D213",
    "D407",
    "D416",
    "E501",
    "ISC001",
    "PLW1514",
    "SIM108",
]
task-tags = ["cspell"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"**/003.ipynb" = ["B023", "PLC2701"]
"**/011.ipynb" = ["PLC2701", "PLR6301"]
"**/014.ipynb" = ["PLC2701"]
"**/016.ipynb" = ["PLC2701"]
"**/021.ipynb" = ["I001"]
"**/022.ipynb" = ["PLC2701"]
"**/024.ipynb" = ["E731", "E741", "S307"]
"**/025.ipynb" = ["E731"]
"**/98*.ipynb" = ["E731", "PLR6301"]
"*.ipynb" = [
    "A003",
    "B008",
    "B018",
    "C408",
    "C90",
    "D",
    "E402",
    "E703",
    "E741",
    "F404",
    "N802",
    "N803",
    "N806",
    "N807",
    "N815",
    "N816",
    "PLC0415",
    "PLC2401",
    "PLR09",
    "PLR2004",
    "PLW0602",
    "PLW0603",
    "PLW2901",
    "RUF027", # for _latex_repr_
    "S101",
    "S307",
    "T20",
    "TCH00",
]
".pre-commit/__init__.py" = ["N999"]
"docs/*" = [
    "D",
    "E402",
    "INP001",
    "S101",
    "S113",
    "T201",
]
"docs/conf.py" = [
    "A001",
    "D100",
    "PLC2701",
    "PLW2901",
]
"docs/report/002.ipynb" = ["F821"]
"docs/report/003.ipynb" = ["F821", "S404", "S603"]
"docs/report/004.ipynb" = ["S404", "S603"]
"docs/report/005.ipynb" = ["F821", "S404", "S603"]
"docs/report/006.ipynb" = ["S404", "S603"]
"docs/report/009.ipynb" = ["S404", "S603"]
"docs/report/010.ipynb" = ["S404", "S603"]
"docs/report/011.ipynb" = ["F821"]
"docs/report/017.ipynb" = ["S404", "S603"]
"docs/report/020.ipynb" = ["F821"]
"docs/report/021.ipynb" = ["S404", "S603"]
"docs/report/022.ipynb" = ["F821"]
"docs/report/024.ipynb" = ["F821", "S102"]
"docs/report/025.ipynb" = ["S404", "S603"]
"docs/report/027.ipynb" = ["PLW1510", "S404", "S603"]
"pin_nb_requirements.py" = ["PLW2901"]
"setup.py" = ["D100"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.tomlsort]
all = false
ignore_case = true
in_place = true
sort_first = [
    "build-system",
    "project",
    "tool.setuptools",
    "tool.setuptools_scm",
]
sort_table_keys = true
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
